Update get_backtest_data.py
Fix bug caused by yfinance changing how they report dates.
Added a bit of boilerplate so that the code can be run in parallel.

This is an ugly hack that leaves a lot of extra directories around and doesn't clean up after itself. However, that doesn't matter in a colab notebook setting.